+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.12.0 / 2014-06-19
+
+* Go 1.3+ required on Windows (uses syscall.ERROR_MORE_DATA internally).
+* [internal] remove cookie from Event struct (unused).
+* [internal] Event struct has the same definition across every OS.
+* [internal] remove internal watch and removeWatch methods.
+
 ## v0.11.0 / 2014-06-12
 
 * [API] Renamed Watch() to Add() and RemoveWatch() to Remove().